pyproject.toml
[project]
name = "automix-ai"
version = "0.1.0"
description = "A Python CLI tool that analyzes audio files to identify optimal DJ mix transition points"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "soundcloud-v2>=1.5.0",
    "ruff>=0.14.14",
    "yt-dlp>=2024.0.0",
]

[project.scripts]
automix = "automix.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "soundfile>=0.12.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["automix"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
